+++ lib/types/private-types.ts
@@ -1,12 +1,4 @@
 /* eslint-disable @typescript-eslint/ban-types */
-/**
- * Extracts the names of all non-function properties of a type.
- */
-export type NonFunctionPropertyNames<T> = {
-  [K in keyof T]: T[K] extends Function ? never : K;
-} extends { [_ in keyof T]: infer U }
-  ? U
-  : never;
 
 export type NonNumberPropertyNames<T> = {
   [K in keyof T]: T[K] extends number ? never : K;